Reorg components & impl. TxRef component
